litellm_project/pyproject.toml
[project]
name = "litellm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.71.2",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
]
